Fix C4189 and C4100 Warnings
C4189 is 'Local variable initialised but not used'
C4100 is 'Unreferenced parameter'

Fixed some other warnings, e.g. about empty if, too.

+++ src/Mix.cpp
@@ -107,7 +107,7 @@ void MixAndRender(TrackList *tracks, TrackFactory *trackFactory,
    }
 
    /* create the destination track (NEW track) */
-   if (numWaves == TrackList::Channels(first).size())
+   if (numWaves == (int)TrackList::Channels(first).size())
       oneinput = true;
    // only one input track (either 1 mono or one linked stereo pair)
